add empty migration
